(PC-33494)[API] feat: remove deprecated api routes (venues' stocks)
