+++ api/src/pcapi/routes/public/blueprints.py
@@ -10,7 +10,6 @@ from . import spectree_schemas
 
 # DEPRECATED APIs
 from .booking_token import blueprint as booking_token_blueprint
-from .books_stocks import blueprint as booking_stocks_blueprint
 
 
 DEPRECATED_PUBLIC_API_URL_PREFIX = "/v2"
@@ -34,7 +33,6 @@ deprecated_v2_prefixed_public_api = Blueprint(
     "deprecated_public_api", __name__, url_prefix=DEPRECATED_PUBLIC_API_URL_PREFIX
 )
 deprecated_v2_prefixed_public_api.register_blueprint(booking_token_blueprint.deprecated_booking_token_blueprint)
-deprecated_v2_prefixed_public_api.register_blueprint(booking_stocks_blueprint.deprecated_books_stocks_blueprint)
 # Deprecated collective endpoints
 deprecated_collective_public_api = Blueprint("public_api_deprecated", __name__, url_prefix="/")
