--- api/src/pcapi/routes/public/books_stocks/__init__.py
@@ -1,6 +1,0 @@
-from flask import Flask
-
-
-def install_routes(app: Flask) -> None:
-    # pylint: disable=unused-import
-    from . import endpoints
