--- api/src/pcapi/routes/public/books_stocks/blueprint.py
@@ -1,4 +1,0 @@
-from flask import Blueprint
-
-
-deprecated_books_stocks_blueprint = Blueprint("books_stocks", __name__)
